Rholang linker: js file IO: complete draft
